requirements.txt
fastapi
uvicorn
sqlalchemy
sqlmodels